+++ lib/src/vector_space_model.dart
@@ -1,0 +1,8 @@
+// Copyright ©2022, GM Consult (Pty) Ltd.
+// BSD 3-Clause License
+// All rights reserved
+
+/// Stub for [VectorSpaceModel].
+class VectorSpaceModel {
+  //TODO: implement class [VectorSpaceModel]
+}