+++ CHANGELOG.md
@@ -4,6 +4,18 @@ Copyright (c) 2022, GM Consult Pty Ltd
 All rights reserved. 
 -->
 
+### 0.0.1-beta.2
+
+- **PRE-RELEASE**
+- Added object model class `FreeTextQuery`.
+- Added interface class `FreeTextSearch`.
+- Added private implementation class `_FreeTextSearchImpl`.
+- Added enum `QueryTermModifier`.
+- Added class `QueryParser`.
+- Added stub for class `SearchResultScorer`.
+- Added object model class `SearchResult`.
+- Added stub for class `VectorSpaceModel`.
+
 ### 0.0.1-beta.1
 
 - **PRE-RELEASE**